registerBot/requirements.txt
botbuilder-integration-aiohttp>=4.11.0
botbuilder-dialogs>=4.11.0
botbuilder-ai>=4.11.0
